+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 CHANGELOG
 =========
 
-5.7, 2024-12-02
+5.7, 2025-03-07
 ---------------
+- add separate button for creating database from Skype export archive (issue #129)
 - fix opening third-party license link in snaps
 - fix error on right-dragging in database list (issue #128)
 - fix startup error on Windows with no English locale available (issue #129)
-- improve compatibility Python 3.10+
+- improve compatibility with Python 3.10+
 - fix selection issues after removing databases from index page
 - fix saving search help text into configuration file